Rename MyExeServerCpp project to MyExeServerAtl.
