+++ MyExeServerAtl/Resource.h
