+++ MyExeServerAtl/Resource.rc
